+++ mirai-core-qqandroid/src/commonMain/kotlin/net/mamoe/mirai/qqandroid/message/MessageSourceFromMsg.kt
@@ -11,6 +11,7 @@ package net.mamoe.mirai.qqandroid.message
 
 import net.mamoe.mirai.contact.Group
 import net.mamoe.mirai.message.data.MessageSource
+import net.mamoe.mirai.message.data.messageRandom
 import net.mamoe.mirai.qqandroid.io.serialization.loadAs
 import net.mamoe.mirai.qqandroid.io.serialization.toByteArray
 import net.mamoe.mirai.qqandroid.network.protocol.data.proto.ImMsgBody
@@ -21,13 +22,16 @@ internal inline class MessageSourceFromServer(
     val delegate: ImMsgBody.SourceMsg
 ) : MessageSource {
     override val time: Long get() = delegate.time.toLong() and 0xFFFFFFFF
-    override val sequenceId: Int get() = delegate.origSeqs?.firstOrNull() ?: error("cannot find sequenceId from ImMsgBody.SourceMsg")
+
+    override val id: Long
+        get() = (delegate.origSeqs?.firstOrNull() ?: error("cannot find sequenceId from ImMsgBody.SourceMsg")).toLong().shl(32) or
+                (delegate.pbReserve.loadAs(SourceMsg.ResvAttr.serializer()).origUids!!.toInt()).toLong().and(0xFFFFFFFF)
+
 
     override suspend fun ensureSequenceIdAvailable() {
         // nothing to do
     }
 
-    override val messageUid: Int get() = delegate.pbReserve.loadAs(SourceMsg.ResvAttr.serializer()).origUids!!.toInt()
     // override val sourceMessage: MessageChain get() = delegate.toMessageChain()
     override val senderId: Long get() = delegate.senderUin
     override val groupId: Long get() = Group.calculateGroupCodeByGroupUin(delegate.toUin)
@@ -39,12 +43,14 @@ internal inline class MessageSourceFromMsg(
     val delegate: MsgComm.Msg
 ) : MessageSource {
     override val time: Long get() = delegate.msgHead.msgTime.toLong() and 0xFFFFFFFF
-    override val sequenceId: Int get() = delegate.msgHead.msgSeq
+    override val id: Long
+        get() = delegate.msgHead.msgSeq.toLong().shl(32) or
+                delegate.msgBody.richText.attr!!.random.toLong().and(0xFFFFFFFF)
+
     override suspend fun ensureSequenceIdAvailable() {
         // nothing to do
     }
 
-    override val messageUid: Int get() = delegate.msgBody.richText.attr!!.random
     // override val sourceMessage: MessageChain get() = delegate.toMessageChain()
     override val senderId: Long get() = delegate.msgHead.fromUin
     override val groupId: Long get() = delegate.msgHead.groupInfo!!.groupCode
@@ -62,7 +68,7 @@ internal inline class MessageSourceFromMsg(
             type = 0,
             time = delegate.msgHead.msgTime,
             pbReserve = SourceMsg.ResvAttr(
-                origUids = messageUid.toLong() and 0xffFFffFF
+                origUids = messageRandom.toLong() and 0xffFFffFF
             ).toByteArray(SourceMsg.ResvAttr.serializer()),
             srcMsg = MsgComm.Msg(
                 msgHead = MsgComm.MsgHead(
@@ -72,7 +78,7 @@ internal inline class MessageSourceFromMsg(
                     c2cCmd = delegate.msgHead.c2cCmd,
                     msgSeq = delegate.msgHead.msgSeq,
                     msgTime = delegate.msgHead.msgTime,
-                    msgUid = messageUid.toLong() and 0xffFFffFF
+                    msgUid = messageRandom.toLong() and 0xffFFffFF
                     , // ok
                     groupInfo = MsgComm.GroupInfo(groupCode = delegate.msgHead.groupInfo.groupCode),
                     isSrcMsg = true